pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "skribe"
version = "0.1.4"
description = "Property testing for Stylus smart contracts"
authors = [
    "Runtime Verification, Inc. <contact@runtimeverification.com>",
]

[tool.poetry.scripts]
skribe = "skribe.__main__:main"
skribe-simulation = "skribe.simulation:main"

[tool.poetry.plugins.kdist]
stylus-semantics = "skribe.kdist.plugin"

[tool.poetry.dependencies]
python = "^3.10"
pykwasm = { git = "https://github.com/runtimeverification/wasm-semantics.git", tag = "v0.1.124", subdirectory = "pykwasm" }
eth-abi = "^5.2.0"
pycryptodome = "^3.22.0"

[tool.poetry.group.dev.dependencies]
autoflake = "*"
black = "*"
flake8 = "*"
flake8-bugbear = "*"
flake8-comprehensions = "*"
flake8-quotes = "*"
flake8-type-checking = "*"
isort = "*"
mypy = "*"
pep8-naming = "*"
pytest = "*"
pytest-cov = "*"
pytest-mock = "*"
pytest-xdist = "*"
pyupgrade = "*"

[tool.isort]
profile = "black"
line_length = 120

[tool.autoflake]
recursive = true
expand-star-imports = true
remove-all-unused-imports = true
ignore-init-module-imports = true
remove-duplicate-keys = true
remove-unused-variables = true

[tool.black]
line-length = 120
skip-string-normalization = true

[tool.mypy]
disallow_untyped_defs = true
